// src/api/schemas/walletListResponseDto.ts
/**
 * Generated by orval v8.0.0 🍺
 * Do not edit manually.
 * Gourmet7 Agent API
 * AI Agent for DeFi operations
 * OpenAPI spec version: 1.0
 */
import type { WalletListResponseDtoDataItem } from './walletListResponseDtoDataItem';
import type { WalletListResponseDtoMeta } from './walletListResponseDtoMeta';

export interface WalletListResponseDto {
  data: WalletListResponseDtoDataItem[];
  meta: WalletListResponseDtoMeta;
}