// src/api/schemas/walletListResponseDtoDataItem.ts
/**
 * Generated by orval v8.0.0 🍺
 * Do not edit manually.
 * Gourmet7 Agent API
 * AI Agent for DeFi operations
 * OpenAPI spec version: 1.0
 */
import type { WalletListResponseDtoDataItemType } from './walletListResponseDtoDataItemType';

export type WalletListResponseDtoDataItem = {
  id: string;
  address: string;
  name: string;
  type: WalletListResponseDtoDataItemType;
  ownerId: string;
  /** @nullable */
  agentId: string | null;
  createdAt: unknown;
  updatedAt: unknown;
};